pyproject.toml
[tool.poetry]
name = "runmini"
version = "0.1.0"
description = ""
authors = ["Aiden-Jeon <aiden.jongseob@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app", from="src"},
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.1"
langchain = "^0.3.7"
langchain-community = "^0.3.5"
langchain-openai = "^0.2.6"
langchain-cli = "^0.0.31"
langchain-chroma = "^0.1.4"
pydantic-settings = "^2.6.1"
streamlit = "^1.40.1"
litellm = "^1.52.15"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
